fix(wcag): Remove redundant breadcrumb on homepage
This commit resolves an issue where a redundant "Home" link was rendered by the breadcrumb on the site's front page. This created unnecessary noise for users of assistive technology and visual clutter.

The `breadcrumb.html.twig` template has been updated to include a conditional check for `not is_front`. The breadcrumb component will now only be rendered if the current page is not the homepage.

This change eliminates the redundant link, streamlining the user experience for all users on the front page.